requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.4
djangorestframework-simplejwt>=5.2.2
psycopg2-binary>=2.9.3
uvicorn>=0.17.6
pytz>=2021.1
